:truck: Move DigiD/eHerkenning OIDC generics to top-level package
