+++ src/digid_eherkenning_oidc_generics/__init__.py
@@ -1,0 +1,3 @@
+default_app_config = (
+    "digid_eherkenning_oidc_generics.apps.DigiDeHerkenningOIDCAppConfig"
+)
