+++ src/openforms/conf/base.py
@@ -159,6 +159,7 @@ INSTALLED_APPS = [
     "stuf.stuf_zds",
     "mozilla_django_oidc",
     "mozilla_django_oidc_db",
+    "digid_eherkenning_oidc_generics",
     "django_filters",
     "csp",
     "cspreports",
